Enable/disable based on preprocessor defines & cleanup
Instead of force disabling in release builds (and in editor when the dev build setting is disabled?) dbg draw is enabled via `DBG_DRAW_ENABLED`.
It can also be force disabled with the `DBG_DRAW_DISABLED` define, which is helpful in situations where it's easy to add new defines but not easy to remove existing defines (eg. some build tools).

Also partial cleanup based on LFG standards.
Automated reformat was run for things like syntax style, but naming fixes will need to be done manually.